Prevent walking around a thousand unneeded directories during buildworld 'includes' phase by excluding directories that have no INCS.
The 'includes' phase is only used to populate the WORLDTMP sysroot used
for later building.  If for some reason a directory that has includes is
removed here then something later will just not build.  Its include
files would otherwise be built during the 'everything' phase and
installed during 'installworld' just fine.  So there is little risk to
doing this except some extra maintenance cost.  It could be argued that
the 2 special cases in usr.bin/ are in the wrong place.

This saves a minute on my system with -j15, which is significant for -DNO_CLEAN
builds which now finish in 2.5 minutes with no changes.

Sponsored by:	EMC / Isilon Storage Division