Optimize 'make depend' for bsd.lib.mk.
There's no need to reprocess the .depend file to add in the .po and .So
files since we can achieve the same thing in bsd.lib.mk.  This now just
depends the .po and .So files on the .o file.  For parallel builds it is
slightly less efficient since it will wait for the main .o file to be
created but this saves wasted time for 'make depend' and avoids a lot of
overhead for directories such as lib/libc that have a large .depend
file.

Sponsored by:	EMC / Isilon Storage Division