bsd.subdir.mk: Connect SUBDIRS to the build regardless of SUBDIR_PARALLEL.
For SUBDIR_PARALLEL a <target>_subdir_<dir> target is made that is
hooked into the real <target>.  For non-SUBDIR_PARALLEL we used a .USE
of _SUBDIR attached to <target>.  For the sake of consistency remove the
use of _SUBDIR for non-SUBDIR_PARALLEL and just build nodes that
maintain the proper ordering and use those.  This keeps the build output
for -j builds to be more consistent as well.  It also allows for fixing
some ordering issues with other targets that need to put SUBDIR
recursing before them, which cannot easily be done with mixed-mode
recursion.

Add some more comments and rearrange a few conditions while here.

MFC after:	never
Sponsored by:	EMC / Isilon Storage Division