Add a way to disallow shlibs with missing symbols for libraries supporting LIBADD.
When not building a module, it is usually proper to ensure all symbols
a library uses is provided by the libraries it links in.  There are
exceptions to this but it is generally true as binaries should not need
to link in libraries they do not use just because some library they do
use is under-linked.  More work is needed to find the exceptions and allow
them to continue building as they do now.  For now add in a way to disable
and enable this for further development.  This is also discussed in the ld(1)
'--no-allow-shlib-undefined' section, which is not quite what we want here
either as it complains about libc not providing environ/__progname, which is
expected, when linking into consumers.  There are also special cases where a
library has a cyclic dependency with another library and cannot provide itself
fully linked to its dependencies.

For many exceptions, replacing the API with weak symbols will allow this
check to pass.  For cyclic dependency cases if the symbols cannot be
easily reworked then combining the libraries is usually the right
approach.

This is to fix under-linked libraries and part of the effort to also
reduce over-linking in binaries.

Sponsored by:	EMC / Isilon Storage Division