Add a way to disallow creating a shared library with missing symbols.
This is only supported for libraries known to LIBADD.

When not building a module or LD_PRELOAD library, it is usually proper
to ensure all symbols a library uses are provided by the libraries it linked
in.  There are exceptions to this but it is generally true.  Binaries should
not need to link in libraries they do not use directly just because some
library they do use is under-linked.  More work is needed to find the
exceptions and allow them to continue building as they do now.  For now add
in a way to disable and enable this for further development.  This is also
discussed in the ld(1) '--no-allow-shlib-undefined' section, which is not
quite what is wanted here either as it complains about libc not providing
'environ'/'__progname', which is expected, when linking into consumers.  There
are also special cases where a library has a cyclic dependency with another
library and cannot provide itself fully linked to its dependencies.

For many exceptions, replacing the API with weak stub symbols will allow
this check to pass.  For cyclic dependency cases if the symbols cannot be
easily reworked then combining the libraries is usually the right
approach.

This is to fix under-linked libraries and part of the effort to also
reduce over-linking in binaries.

LINKTEST is chosen and not yet added to src.opts.mk since there is
future work coming to check for over-linking in binaries, in the build,
using the tools/builds/check-links.sh script.  At that time the option
will graduate.

Sponsored by:	EMC / Isilon Storage Division