Use 'all-install' for most of 'make libraries'.
This reduces the need to have multiple stages of libraries and is needed
to connect contrib/ofed/usr.lib after fixing underlinking there.

This also allows the 'make install' phase for lib/ into WORLDTMP to be
parallel, which is normally disallowed in lib/Makefile due to safety of a
running system.

Sponsored by:	EMC / Isilon Storage Division