Allow libnv to be built externally using GCC.
GCC does not define _VA_LIST_DECLARED.  It defines _VA_LIST_ and others.
This was causing the prototype to not be defined and leading to an error
later due to using nvlist_add_stringv(3) without a prototype in
nvlist_add_stringf(3).

This uses the same check as other va_list prototypes in the original
change in r279438.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@293361 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f