Use some bmake features to delay execution until actually needed.
This avoids executing these during tree-walks that would not
otherwise need the variables.

Note that '$$<var>' inside of the :L:sh trick does not expand as expected so
is avoided.

Note also that these will exec on every expansion so are not a
productive method for loops.

Sponsored by:	EMC / Isilon Storage Division