FIXME XXX: DIRDEPS_BUILD: Stop tracking 'local dependencies'.
These were only in Makefile.depend for 'clean builds' since nothing would
generate the files on an incremental build.  Normally if a .depend file
does not exist then a dependency is added by bsd.lib.mk/bsd.prog.mk
from OBJS: *.h.  However, meta.autodep.mk creates a .depend file from
created meta files and cats that into Makefile.depend.  It only
tracks *.[ch] files though which is basically forcing 'make depend'
(a dependency on ${SRCS}) to run when starting a build.  This .depend
that is created then breaks incremental builds though.  The goal was
to skip 'make depend' yet it only really works the first time.  After
that files are not generated as expected, which r288966 tried to address
but was using buildfiles: rather than beforebuild: and was reverted
in r291725.  This was resolved in XXX though which always depends
on SRCS before starting the build.

meta.autodep.mk is still used to generate Makefile.depend though via:
  gendirdeps: Makefile.depend
  .END: gendirdeps

This commit allows removing all of the local dependencies in
Makefile.depend which cuts down on churn and some of the arch-dependent
changes in the files.

Sponsored by:	EMC / Isilon Storage Division