Begin abstracting out the access method for ath(4) ioctls.
Each of the ath* commands do their own direct socket/ioctl calls,
which makes it difficult to forklift upgrade things.

So, this is the beginning of abstracting out the stats API calls
in the hope that I can migrate things to use a /dev/athX file for
ioctls and use a more general interface.

Tested:

* QCA9565 NIC, STA mode


git-svn-id: svn+ssh://svn.freebsd.org/base/head@296152 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f