Add a subdir-${.TARGET} hook and fix 'make all' to recurse before building.
This could be done by attaching the subdir targets to beforebuild: but
by allowing a 'subdir-${TARGET}' attachment it allows placing the
recursion exactly where wanted in all cases.

'all:' is special because so much attaches to it in share/mk, while the
subdir attachment comes last.  Other targets such as 'obj' and 'clean'
were already recursing first so this just makes things consistent.
'beforebuild' is ran before recursing here to be consistent with
'beforeinstall' doing the same and giving a chance to build something in
the directory before recursing.

MFC after:	2 weeks
Sponsored by:	EMC / Isilon Storage Division