Add STATIC_LIBS option to allow disabling build/install of optional static archives.
This is useful for systems that don't need static support and are tight
on space.

It is not redundant with defining SHLIB_NAME with LIB since this feature will
apply to all builds, not just ones that use LIB.  It is close to what
WITHOUT_INSTALLLIB is but better since it will not build the optional libraries
either and is not broken during build.  INSTALLLIB should not even be a user
option at this point, but it is still useful for the build of libpam static
modules for which was originall added for in r96512.

Internal libraries are still built.

DPADD is ignored for non-internallibs in this case as there will no file to
track.

Sponsored by:	EMC / Isilon Storage Division