Add STATIC_LIBS option to allow disabling build/install of optional static archives.
This is useful for systems that don't need static support and are tight
on space.  Libraries that do not provide a shared library (by setting
NO_PIC) will still build and install a static library.  Internal libraries
continue to function as before.

DPADD is ignored for non-internallibs in this case as there will no file to
track.

It is not redundant with defining SHLIB_NAME with LIB since this feature will
apply to all builds, not just ones that use LIB.

It is close to what WITHOUT_INSTALLLIB is but better since it is not broken by
ensuring lonely static libraries such as libgcc are still installed.  Most of
these are critical for compiling.  INSTALLLIB was originally added in r96512
(and is still useful for) building static libpam modules but not installing
them since they are linked into the static libpam.  The INSTALLLIB option is
not very useful overall since it creates a system which cannot build, nor can
it be used during the build.  It could easily be replaced by users with a "find
${DESTDIR} -name '*.a' -delete".

Sponsored by:	EMC / Isilon Storage Division
Differential Revision:	https://reviews.freebsd.org/D5953