NFS remove: Send the remove RPC when a vnode is in-use with more than 1 link.
Put another way, only do the silly rename on a file in-use if it has 1 link
count and hasn't been renamed yet.

Before this if a vnode had more than 1 link and the directory it was in was
removed with 'rm -rf', then the silly rename would rename the file and the
'rm' would fail since FTS didn't know about the newly renamed file and would
not remove the .nfs file before trying the rmdir().  This is still the case
when there is only 1 link on an in-use vnode, but now if there is more than 1
link then the file is not renamed and the remove RPC is sent, thus the 'rm
-rf' can succeed.

This manifested by hardlinking an executable from a temporary directory into
/, and then running the / version while removing the temporary version.

MFC after:	2 weeks
Sponsored by:	EMC / Isilon Storage Division
Differential Revision:	https://reviews.freebsd.org/D6254