thr_*(2): Add xrefs to what libthr implements using each syscall.
Add text to thr_exit(2) and thr_new(2) discouraging their use in
applications since calling these in a process with libthr loaded will
confuse libthr and is likely to cause hangs or crashes.

The thr_kill2(2) call is not used by libthr and may be useful in special
applications.

The other calls can be used in applications but it should not be necessary.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@301171 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f