WITH_SYSTEM_COMPILER: Disable with WITH_META_MODE.
The WITH_META_MODE build is intended to be a working incremental build.
It spies on the build command to see if things should be rebuilt if the
command changes.  If you run buildworld, it builds a cross-compiler,
then do installworld and buildworld again it will invoke the
WITH_SYSTEM_COMPILER logic.  This then adds on -target/--sysroot, etc,
and causes rebuilds due to the changed build command.  Since the
incremental build is not cleaning anything by default then there is much
less risk to rebuilding the already-existing cross-compiler.  Just
disable the combined logic and always use and build the cross-compiler.

An alternative to this would be to always pass -target/--sysroot.  A
problem with not taking the approach in this commit is that after an
installworld the next build would see a new /usr/bin/cc timestamp
compared to the built-objects and cause them to rebuild.  There's no
good way to fix that otherwise since we clearly want to rebuild if the
compiler changes.

Sponsored by:	EMC / Isilon Storage Division