WITH_META_MODE: Avoid host tool timestamps causing a rebuild.
Using buildworld, installworld, buildworld.  It is expected that nothing
should rebuild.  However any host tool used will have its timestamp
updated.  Any library used by dynamic tools will have its timestamp
updated.  The filemon(4) data in the .meta files captures all reads to
these files.  This causes the 2nd buildworld to rebuild everything since
host tools and files have been updated.  Mitigate this problem by
ignoring tools and system libraries that are unlikely to impact the output
of the build.  In the system library case it is possible they get some
fix to change the output but are likely to be detected by the tool using
them anyhow.  In the case of build-tools/bootstrap-tools/cross-tools it
is safe to ignore all the host compiler and headers since these phases
are already self-reliant.  Not ignoring them would cause them to rebuild
due to changed /usr/bin/cc and changed headers, thus causing a chain
reaction of rebuilds due to the cross-tools being rebuilt.

Longterm this may be problematic for maintenance but otherwise
WITH_META_MODE loses its usefulness for buildworld->installworld uses.
It is otherwise not generally affected by this problem for cases of
building images.

The risk to this change is that it may ignore a change to a file used
in the build that would actually change the output of the build, and
thus break the "working incremental build" guarantee.  Thus any changes
to these lists must be very carefully considered.

An alternative solution would be to update the mtime of all files in the
object directory after installworld so that the host files are not newer
than the object files.

Reported by:	many
Sponsored by:	EMC / Isilon Storage Division