Don't forget to set sa->narg for CloudABI system calls.
It turns out that this value is not used within the system call code
under normal conditions, except when using tracing tools like ktrace.
If we forget to set this value, it is set to random garbage. This may
cause ktrace to hang indefinitely, making it impossible to kill.

Reported by: Michael Plass
PR: 210800
MFC before: 11.0-RELEASE


git-svn-id: svn+ssh://svn.freebsd.org/base/head@302448 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f