pf: Map hook returns onto the correct error values
pf returns PF_PASS, PF_DROP, ... in the netpfil hooks, but the hook callers
expect to get E<foo> error codes.
Map the returns values. A pass is 0 (everything is OK), anything else means
pf ate the packet, so return EACCES, which tells the stack not to emit an ICMP
error message.

PR:	207598


git-svn-id: svn+ssh://svn.freebsd.org/base/head@302497 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f