Change the return type of msgrcv() to ssize_t as required by POSIX.
It looks like the msgrcv() system call is already written in such a way
that the size is internally computed as a size_t and written into all of
td_retval[0]. This means that it is effectively already returning
ssize_t. It's just that the userspace prototype doesn't match up.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@303435 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f