cam_periph_ccbwait could return while ccb in progress
Immediately after cam_ccbq_remove_ccb, a transient state exists where
the ccb index had transitioned from a ccbq index (> 0) but not yet to
the intended future index (normally CAM_ACTIVE_INDEX but possibly
CAM_DONEQ_INDEX during an abort).  During that time it would instead
have the index CAM_UNQUEUED_INDEX, which, alas, is also used to
indicate that a ccb is done.  If cam_periph_ccbwait encountered the ccb
during this window, it was possible for it to interpret the ccb as
being done (despite a check of the status, due to the incoherent nature
of the two reads).  Now, an unambiguous index value represents this
transient state.