libproc: Cache symbol tables for mapped objects upon access.
Extend the file handle cache entries to include symbol tables as well. An
index is used to implement binary search by symbol value. Lookups by
name are comparatively rare and are thus still implemented with a linear
search, but support for a binary search by name would be straightforward
to add if needed.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@309596 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f