Increase interrupt cells in generic_pcie_fdt_route_interrupt
ARM GIC specification in device trees use 3 cells, so the current
limit of 2 causes the last cell to be dropped. This in turn can
cause the interrupt polarity and trigger settings to be incorrect.

Increase the limit to 4 which should handle all reasonable cases.
This fixes issues seen in QEMU when registering PCI interrupts.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@309890 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f