hwpmc: partially depessimize munmap handling if the module is not loaded
HWPMC_HOOKS is enabled in GENERIC and triggers some work avoidable in the
common (module not loaded) case.

In particular this avoids permission checks + lock downgrade
singlethreaded and in cases were an executable mapping is found the pmc
sx lock is no longer bounced.

Note this is a band aid.

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@312724 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f