Enable usb low and full speed devices connected to the imx6 root hubs.
This enables the PHY circuitry for UTMI+ level 2 and 3, and sets the
flag to tell the ehci code that the root hub has a transaction translator
in it.  For imx6 we can use the standard ehci_get_port_speed_portsc()
function to find out what speed device is connected to the port.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@313674 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f