For MD_PRELOAD type md(4) devices, if there is a file name in the preloaded meta-data, copy it into the softc structure.
When returning md(4) device details to the caller, include the file name in
any MD_PRELOAD type devices if it is set (first character is not NUL.)

In mdconfig, for "preload" type md(4) devices, if there is file config
available, print it in the file column of the output.

Reviewed by:	brooks
Approved by:	sjg (mentor)
MFC after:	1 month
Sponsored by:	Juniper Networks, Inc.
Differential Revision:	https://reviews.freebsd.org/D9529


git-svn-id: svn+ssh://svn.freebsd.org/base/head@313701 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f