Fix the attribute for scteken_clear() (change it back from the user user default normal attribute to the current attribute).
This change only fixes a logic error.  scterm_clear() used to be
used for terminal reset, but teken uses a general fill function for
that, leaving scterm_clear() only used for initialization and mode
change, when using the user default attribute is correct.  It is not
really a terminal function, but needs to sync its changes with the
terminal layer.  Syncing of the attribute is currently broken for
terminal reset, but works for initialization and mode change.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@315066 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f