cxgbe/t4_tom: Per-connection rate limiting for TCP sockets handled by the TOE.  For now this capability is always enabled in kernels with options RATELIMIT.  t4_tom will check if_capenable once the base driver gets code to support rate limiting for any socket (TOE or not).
This was tested with iperf3 and netperf ToT as they already support
SO_MAX_PACING_RATE sockopt.  There is a bug in firmwares prior to
1.16.45.0 that affects the BSD driver only and results in rate-limiting
at an incorrect rate.  This will resolve by itself as soon as 1.16.45.0
or later firmware shows up in the driver.

Relnotes:	Yes
Sponsored by:	Chelsio Communications


git-svn-id: svn+ssh://svn.freebsd.org/base/head@317849 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f