Default to using more efficient 'cleanworld' rather than traversing with 'cleandir'.
This moves '_objclean' to the first stage of 'buildworld' so that now calling
'cleanworld' will not remove the newly-populated WORLDTMP since it is removing
the entire obj directory.

This also avoids ESTALE errors for NFS due to not traversing most of
the time.

Also fix 'cleanworld' to use WMAKE and handle libcompat directories for
its fallback cleandir call and respect CLEANDIR.

MFC after:	1 month
Sponsored by:	Dell EMC Isilon