Replace the checks for MNTK_UNMOUNTF with a macro that does the same thing.
This patch defines a macro that checks for MNTK_UNMOUNTF and replaces
explicit checks with this macro. It has no effect on semantics, but
prepares the code for a future patch where there will also be a
NFS specific flag for "forced dismount about to occur".

Suggested by:	kib
MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@321628 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f