Bugfixes and enhancements...
Don't enable the oscillator when it is found to be stopped at init time,
just let the first setting of valid time start it.  But still report a dead
battery if it's stopped at init time.

Don't force the chip into 24hr mode, just cope with whatever mode it is
already in.

Schedule the clock_settime() callbacks to align the RTC clock to top of
second when setting it.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@321823 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f