Remove the ds133x and s35390a i2c RTC drivers for now.  They both do i2c transfers in their probe() or attach() routines, and that doesn't work when the low-level controller requires interrupts to be functional.
The DS133x family of chips is nearly identical to the DS1307 and support
for them should be added to that driver, then the ds133x driver can be
deleted.  The s35390a driver just needs a non-trivial workover.  In both
cases that work will be done and committed separately.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@322282 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f