Rename curr_curs_attr to base_curr_attr.  The actual current cursor attribute field is curs_attr.  The base field holds user data translated in a reversible way and is needed because current field holds this in an irreversible way for efficiency.
Factor out some common code for the reversible translation.  This is
slightly simpler now, and much easier to expand.

Translate the magic flags value -1 to a single control flag internally
up front so other flags can be trusted later.  This can be used for the
relevant ioctl() too.

Remove CONS_CURSOR_FLAGS which contained all the control flags.  It was
unused and not useful.  After adding more flags, there will be tests on
a couple at a time but never on them all.  This API should have used this
to disallow unknown flags.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@322708 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f