Introduce libcaspermock
Summary:
Introduce libcaspermock. This library provides API exactly the same as libcasper. The different is that libcaspermock don't provide sandbox-friendly
functions. Adding this library allow us to simplify the code, we don't need check if libcasper is available in the system anymore, instead we just need to
use this library.

Reviewers: emaste, cem, allanjude, bapt, ed

Reviewed By: ed

Subscribers: bdrewery

Tags: #capsicum

Differential Revision: https://reviews.freebsd.org/D8753