Add notes about overlapping copies.
Add notes to each of these that specifically state that results are
undefined if the strings overlap. In the case of memcpy, we document
the overlapping behavior on FreeBSD (pre-existing). For str*, it is
left unspecified, however, since the default (and x86) implementations
do not handle overlapping strings properly.

PR: 223653
Sponsored by: Netflix


git-svn-id: svn+ssh://svn.freebsd.org/base/head@325765 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f