Complete support for dtrace's -x setenv option.
This allows one to override the environment for processes created with
dtrace -c. By default, the environment is inherited.

This support was originally merged from illumos in r249367 but was lost
when the commit was later reverted and then brought back piecemeal.

Reported by:	Samuel Lepetit <slepetit@apple.com>
MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@326499 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f