awk(1): Add necessary bits for connecting tests, but leave disconnected
The NetBSD test suite has 24 tests for awk, and we pass exactly 4 of them.
Add the necessary pieces for interested parties to easily connect the
tests and run them, but leave them disconnected for the time being.

Some of these tests outright segfault in our awk, others just exhibit the
wrong behavior.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@327806 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f