liblua: Add loader.interpret
This allows lua to pass back a command string to be executed as if it were
typed at the loader prompt- loader tries to execute the string first as pure
lua, then parses it and gives lua a chance to intercept before it tries to
execute it itself.

This will be used to implement menu_timeout_command, among other things,
which *should* be used to execute basic loader commands independent of the
chosen interpreter.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@330339 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f