zfs test suite: align zfs_destroy_005_neg: with upstream
The change is to account for a different order in which the recursive
destroy may be attempted.  If we first try a dataset that can be destroyed
then it will be destroyed, but if we first try a dataset that cannot be
destroyed then we will not attempt to destroy the other dataset.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@330993 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f