zfs: fix mismatch between format specifier and type
vdev_dbgmsg_print_tree printed vdev_id of uint64_t type with %u format
specifier.  That caused subsequent parameters to be incorrectly read
from the stack and lead to a crash when a wrong value was interpreted as
a string pointer.

This should be upstreamed.

Reported by:	pho
MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@331420 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f