In addition to the existing argument format:
	prtblknos filesystem_device inode ...

add an additional argument format:

	prtblknos file

which is more convenient than figuring out the filesystem
and inode number for "file".

When given a list of multiple inodes, rather than exiting
the program on an error with one of them, skip over it and
continue with the next one.

Submitted by: bde


git-svn-id: svn+ssh://svn.freebsd.org/base/head@332742 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f