Break out the cpu_features setup to its own function, to be run earlier
The new POWER9 MMU configuration is slightly different from current setups.
Rather than special-casing on POWER9, move the initialization of cpu_features
and cpu_features2 to as early as possible, so that platform and MMU
configuration can be based upon CPU features instead of specific CPUs if at all
possible.

Reviewed by:	nwhitehorn


git-svn-id: svn+ssh://svn.freebsd.org/base/head@333272 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f