igb(4): I210 restore functionality if pxeboot rom is enabled on this device.
r333345 attempted to determine if this code was needed or it was some kind
of work around for a problem.  Turns out, its definitely a work around for
hardware locking and synchronization that manifests itself if the option
Rom is enabled and is selected as a boot device (there was a PXE attempt).

Reviewed by:	mmacy
Differential Revision:	https://reviews.freebsd.org/D15439


git-svn-id: svn+ssh://svn.freebsd.org/base/head@333631 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f