indent(1): improve handling of boxed comments indentation
The trick is to copy everything from the start of the line into the buffer
that stores newlines and comments until indent finds a brace or an else.
pr_comment() will use that information to calculate the original indentation
of the boxed comment.

This requires storing two pieces of information: the real start of the
buffer (sc_buf) and the start of the comment (save_com).


git-svn-id: svn+ssh://svn.freebsd.org/base/head@334563 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f