Rework WITHOUT_LLD/TOOLCHAIN fix from r327892 for cross-tools.
Also correct an invalid check on MK_LLD_IS_LD for determining which
sources to link into libllvm.  If lld is being built at all these
sources are needed.  Having a link is irrelevant.

MK_LLD is for the installed lld while MK_LLD_BOOTSTRAP is for the build
tool.  For WITH_SYSTEM_LINKER it is necesarry to separate the logic of
these two.  When building libllvm TOOLS_PREFIX will be defined and
MK_LLD_BOOTSTRAP should be checked instead.

Sponsored by:	Dell EMC
Differential Revision:	https://reviews.freebsd.org/D15837