Stop building intermediate .o files.
These are not used to link the final tool anymore.  At some point
in the past the suffix rules changed to not link these in.  The
original reason for this in r19176 is unclear but seems to be
related to mkdep.  The .depend handling is still broken here.

Sponsored by:	Dell EMC