Restore the order in which RFC 3164 messages with fac/pri are formatted.
The refactoring of the syslogd code to format messages using iovecs
slightly altered the output of syslogd by placing the facility/priority
after the hostname, as opposed to printing it right before. This change
reverts the behaviour to be consistent with how it was before.

PR:		229457
Reported by:	Andre Albsmeier
MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@335861 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f