lualoader: Fix (add) Xen support
lualoader was not respecting the 'xen_kernel' environment variable, which
hints to the interpreter that it should load a Xen kernel prior to loading
any other kernel that might be specified. If a Xen kernel is specified and
we fail to load it, we should not proceed to boot.

Reported by:	royger
Tested by:	royger


git-svn-id: svn+ssh://svn.freebsd.org/base/head@338255 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f