lualoader: fix color usage
Resetting to the default color scheme was done prior to reading the config.
This is bogus; colors may only be declined by the user with the
loader.conf(5) variable "loader_color", so such a request for no color will
not be completely honored as we reset to the default color scheme
unconditionally.

Approved by:	re (rgrimes)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@338394 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f