Address the fact that errors from `copyout(9)` are not properly percolated up on failure
Not properly reporting this error can result in values "returned" from
`sbytes` being misinterpreted.

When reporting `sbytes`, be sure to not squash existing errors outside of
EAGAIN, EBUSY, and EINTR. These errors are non-fatal, so they should only be
percolated up on failure if EFAULT isn't reported, as EFAULT affects the
results of `sbytes`.

Bug:	232210