Add the check that current VNET is ready and access to srchash is allowed.
This change is similar to r339646. The callback that checks for appearing
and disappearing of tunnel ingress address can be called during VNET
teardown. To prevent access to already freed memory, add check to the
callback and epoch_wait() call to be sure that callback has finished its
work.

MFC after:	20 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@339649 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f