Add a style.mdoc(5) manual page.
The aim of this manual page is to act as a style and formatting guide for
mdoc(7) manual pages. Currently, mdoc(7) does not provide much guidance
when it comes to the usage of macros making it difficult to format manual
pages in a consistent way.

Reviewed by:	bcr
Approved by:	bcr (doc), krion (mentor, implicit), mat (mentor, implicit)
Differential Revision:	https://reviews.freebsd.org/D18394


git-svn-id: svn+ssh://svn.freebsd.org/base/head@342578 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f