loader: really fix cd9660 dirmatch
The cd9660_open() does pass whole path to dirmatch() and we need to
compare only the current path component, not full path.

Additinally, skip over duplicate / (if any) and check if the last component
in the path was meant to be directory (having trailing /). If it is in fact
a file, error out.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@344387 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f