tuntap: Improve style
No functional change.

tun_flags of the tuntap_driver was renamed to ident_flags to reflect the
fact that it's a subset of the tun_flags that identifies a tuntap device.
This maps more easily (visually) to the TUN_DRIVER_IDENT_MASK that masks off
the bits of tun_flags that are applicable to tuntap driver ident. This is a
purely cosmetic change.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@347483 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f