libbe(3): restructure be_mount, skip canmount check for BE dataset
Further cleanup after r349380; loader and kernel will both ignore canmount
on the root dataset as well, so we should not be so strict about it when
mounting it. be_mount is restructured to make it more clear that depth==0 is
special, and to not try fetching these properties that we won't care about.

MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@349383 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f