In nxprtc(4), use the countdown timer for better timekeeping resolution on PCx2129 chips too.
The datasheet for the PCx2129 chips says that there is only a watchdog
timer, no countdown timer.  It turns out the countdown timer hardware is
there and works just the same as it does on a PCx2127 chip, except that you
can't use it to trigger an interrupt or toggle an output pin.  We don't need
interrupts or output pins, we only need to read the timer register to get
sub-second resolution.  So start treating the 2129 chips the same as 2127.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@350016 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f