Add an initial RELNOTES file.
The intent is to provide a convenient location to document changes
that are relevant to users of binary FreeBSD distributions, in contrast
with UPDATING, which exists to document caveats for users who build
FreeBSD from source.

This complements the "Relnotes:" tag in commit messages by providing a
place to document the change in more detail, or in case a "Relnotes:"
tag was accidentally omitted.  In particular, "Relnotes:" should be
used if you do not intend to document the change in RELNOTES for some
reason.

Changes to the file should not be MFCed.  For now the file will exist
only in head, but may be updated via direct commits to stable branches
depending on how things go.

I took the liberty of pre-populating the file with some recent release
notes-worthy changes.

Reviewed by:	ian (earlier version)
Discussed with:	cy, gjb, imp, rgrimes
Differential Revision:	https://reviews.freebsd.org/D20762


git-svn-id: svn+ssh://svn.freebsd.org/base/head@350089 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f