arm64: Implement HWCAP
Add HWCAP support for arm64.
defines are the same as in Linux and a userland program can use
elf_aux_info to get the data.
We only save the common denominator for all cores in case the
big and little cluster have different support (this is known to
exists even if we don't support those SoCs in FreeBSD)

Differential Revision:	https://reviews.freebsd.org/D17137


git-svn-id: svn+ssh://svn.freebsd.org/base/head@350166 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f