Remove META_MODE whitelist, reverting r301887 and r298220
This never really made sense. If an install target is making a cookie
then it will NOT rebuild during install with META_MODE off. Actually
when META_MODE is enabled and filemon(4) is used then the target
may actually reinstall, even with a cookie, if any of the installed
files are missing or outdated. So it makes more sense to NOT use a
whitelist like this.

The META_DEPS feature allows adding a cookie to a target to prevent
rerun when filemon(4) is used.

The first problem this was addressing was related to 'make delete-old-libs'
not writing its rm -i prompts to stderr with META_MODE enabled.

The second problem was requiring filemon(4) during install where it was
not built at. But as discussed earlier it may make sense to have filemon
support there to handle any cookies improperly used. So this problem is
really about the default/availability of filemon(4).