powerpc/powernv: Only clear EEH freeze for some errors
Only clear an EEH freeze if an error occurs.  However, if an OPAL_HARDWARE
error is returned, this indicates a hardware failure which cannot be
unfrozen, and instead needs a hardware reset.  Attempting to unfreeze a
broken PCH will result in console spam for each attempt.  To avoid the spam,
just don't do it.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@350494 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f