exit1: postpone clearing P_TRACED flag until the proctree lock is acquired
In case of the process being debugged. The P_TRACED is cleared very early,
which would make procdesc_close() not calling proc_clear_orphan().
That would result in the debugged process can not be able to collect
status of the process with process descriptor.

Reviewed by:	markj, kib
Tested by:	pho
MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@350610 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f