Revert r352557: powerpc/loader: Install ubldr without stripping
This was committed due to what was later diagnosed as an msdosfs bug
preventing in-place strip. This bug was fixed in r352564, and we agreed to
keep the workaround in for a bit to allow the driver fix a suitable amount
of propagation time for folks building/installing powerpc/ubldr, seeing as
how we were not in any hurry to revert.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@353161 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f