powerpcspe: Don't leak kernel registers in SPE dumps
save_vec_int() for SPE saves off only the high word of the register, leaving
the low word as "garbage", but really containing whatever was in the kernel
register at the time.  This leaks into core dumps, and in a near future
commit also into ptrace.  Instead, save the GPR in the low word in
save_vec_nodrop(), which is used only for core dumps and ptrace.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@354776 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f