sh: Test that executing various binary files is rejected
If executing a file fails with an [ENOEXEC] error, the shell executes the
file as a shell script, except that this execution may instead result in an
error message if the file is binary.

Per a recent Austin Group interpretation, we will need to change this to
allow a concatenation of a shell script and a binary payload. See
Austin Group bugs #1226 and #1250.

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@356208 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f