Remove prereq lib support for GNU unwinder
LLVM's libunwind is used for all archs, as of r356514.  As GCC 4.2.1
will soon be removed, and with it the GNU unwinder, start simplifying
the prereq/startup lib logic in Makefile.inc1.  From here if there are
any unwinder bugs (on any arch) the path foward is to fix LLVM's
libunwind.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@356516 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f