arm: allwinner: Add GPIO Interrupt support
Not all pins in Allwinner have interrupts support so we rely
on the padconf data to add the proper caps when pin_getcaps is called.
The pin is switch to the specific "eint" function during setup_intr and
switched back to its old function in teardown_intr.
Only INTR_MAP_DATA_GPIO is supported for now.

MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@356891 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f