libc: provide fputc_unlocked
Among the same justification as the other stdio _unlocked; in addition to an
inline version in <stdio.h>, we must provide a function in libc as well for
the functionality. This fixes the lang/gcc* builds, which want to use the
symbol from libc.

PR:		243810
Reported by:	antoine, swills, Michael <michael.adm gmail com>
X-MFC-With:	r357284


git-svn-id: svn+ssh://svn.freebsd.org/base/head@357419 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f