Enable vm_object_mightbedirty() and vm_object_page_clean() for swap objects backing tmpfs vnodes data.
The clean scan is limited to only remove write permissions from the
mapped pages of the objects.  This fixes the issue that tmpfs vnode
mtime is not updated from writes to the mmaped area after the initial
page-in.

Noted by:	mjg
Reviewed by:	markj
Discussed with:	jeff
Tested by:	pho
Sponsored by:	The FreeBSD Foundation
MFC after:	1 week
Differential revision:	https://reviews.freebsd.org/D23432


git-svn-id: svn+ssh://svn.freebsd.org/base/head@357514 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f