O_SEARCH test: drop O_SEARCH|O_RDWR local diff
In FreeBSD's O_SEARCH implementation, O_SEARCH in conjunction with O_RDWR or
O_WRONLY is explicitly rejected. In this case, O_RDWR was not necessary
anyways as the file will get created with or without it.

This was submitted upstream as misc/54940 and committed in rev 1.8 of the
file.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@357580 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f