Move smbios.c to libsa.
smbios used to be an i386 only kinda weird quirk to the x86
architecture. But UEFI picked it up, dusted it off and now it's many
other locations. Make it base technology by moving it to libsa and
fixing up the compliation. The code has issues with unaligned access
still, but that will be addressed in a followup commit.

Differential Revision: https://reviews.freebsd.org/D23660


git-svn-id: svn+ssh://svn.freebsd.org/base/head@358136 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f