Create ../compile
Give up the battle to keep extra files in $MACHINE/compile to keep the file in
the tree. Instead, create CDIR (usually ../compile) if it doesn't exist when
we're using a default build location (eg, not using -d). If it does, we do
nothing. This only affects people that do old-school builds, but it's bit me a
dozen times since last summer so time to fix the bug.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@358291 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f