test if port does exist via using scratch register
The SCR, scratch register was not present on the 8250 and 8250B UART, so we
can use to test if we actually do have serial port.

We need this test because some systems will get long delays while attempting
to write to non-existing port and this will slow down the console IO
to extreme.

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@358906 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f