gdb: compile with -fcommon explicitly
As described in the comment, gdb relies on some of the linker magic that
happens with -fcommon. I suspect the life expectancy of gdb-in-base is low
enough that this isn't worth spending much time addressing, especially given
the vintage. Hit it with the -fcommon hammer so that it continues to just
work.

MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@359432 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f