lualoader: config: start exporting readConfFiles
In the process, change it slightly: readConfFiles will take a string like
loader_conf_files in addition to the loaded_files table that it normally
takes. This is to facilitate the addition of a read-conf CLI command, which
will just pass in the single file to read and an empty table.

MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@360421 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f