Flip kern.tty_info_kstacks on by default
It's a useful debug aid for anyone using Ctrl-T today, and doesn't seem to be
widely known.  So, enable it out of the box to help people find it.

It's a tunable and sysctl, so if you don't like it, it's easy to disable
locally.

If people really hate it, we can always flip it back.

Reported by:	Daniel O'Connor


git-svn-id: svn+ssh://svn.freebsd.org/base/head@362141 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f