Fix the NFSv4 client so that it checks for support of TimeCreate before trying to set it.
r362490 added support for setting of the TimeCreate (va_birthtime) attribute,
but it does so without checking to see if the server supports the attribute.
This could result in NFSERR_ATTRNOTSUPP error replies to the Setattr operation.
This patch adds code to check that the server supports TimeCreate before
attempting to do a Setattr of it to avoid these error returns.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@363587 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f