Have script accept and ignore -e for Linux compat
In the util-linux version of script, it will always exit with succes.
Except when run with -e, in which case it will have the exit value of
the child. BSD Script already uses the child's exit value for its exit
value. Some config and other helper scripts depend on being able to
specify -e. Accept it for compatibility since we'll already to the
right thing, but otherwise we ignore it.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@365052 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f