mtree -c { -O | -X }: Avoid silently ignoring errors from fts(3)
With an only (-O), or excludes (-X), filter it is possible for fts_read(3)
to encounter some error that sets FTS_STOP while looking at an ignored
path.  No error is printed because it is not a wanted path.  Further reads
return NULL indicating a successful end.  Thus later encountered paths
that are expected never are recorded and no error is indicated on stderr
nor for exit status.

While here don't handle jflag logic if this path is ignored.

MFC after:	2 weeks
Sponsored by:	Dell EMC
Differential Revision:	https://reviews.freebsd.org/D27163