Add '-u' switch that would uncompress cores that were compressed by kernel during dump time.
A real life scenario is that cores are compressed to reduce
size of dumpon partition, but we either don't care about space
in the /var/crash or we have a filesystem level compression of
/var/crash. And we want cores to be uncompressed in /var/crash
because we'd like to instantily read them with kgdb. In this
case we want kernel to write cores compressed, but savecore(1)
write them uncompressed.

Reviewed by:	markj, gallatin
Relnotes:	yes
Differential Revision:	https://reviews.freebsd.org/D27245


git-svn-id: svn+ssh://svn.freebsd.org/base/head@367824 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f