Remove concept of mbox_sleep_ok.
It was broken by design and unused for years due to conflicts between
different threads, fighting for the same set of mailbox registers, not
designed for multiple requests at a time.  So either request has to be
synchronous and spin under the lock, or it should be sent asynchronously
through the queues as Mailbox Command IOCB or some other way.

This removes any OS specifics from the wait code, so it can be inlined.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@367986 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f