Make generated C files depend on this Makefile
The contents of lib.c, lib2.c, bc_help.c, and dc_help.c depends on the
parameters passed to strgen.sh in this Makefile. A change to the number
of parameters of strgen.sh has been applied to the invocation of this
command, but this did not cause a rebuild of the generated files.

Reported by:	Cy.Schubert@cschubert.com


git-svn-id: svn+ssh://svn.freebsd.org/base/head@368097 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f