loader: Print autoboot countdown immediately, not at 9
For the first second otime and ntime are equal so no message gets
printed. Instead we should print the countdown right from the start,
although we do it at the end of the first iteration so that if a key has
already been pressed then the message is suppressed.

Reviewed by:	imp
Approved by:	imp
Differential Revision:	https://reviews.freebsd.org/D26935


git-svn-id: svn+ssh://svn.freebsd.org/base/head@368623 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f